start process of moving protocol and pipepline interface classes back to looper
